requirements.txt
rembg[cli]
pillow
onnxruntime
tqdm
